# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: AI Smetchik Backend API v1
  version: 1.0.0
  description: |
    MVP API namespace for split frontend/backend deployment.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /v1/auth/login:
    post:
      security: []
      summary: Login and issue JWT tokens
      requestBody:
        required: true
      responses:
        '200':
          description: OK
...